\name{NAProbes}
\alias{NAProbes}
\alias{print.NAProbes}
\alias{summary.NAProbes}
\alias{plot.NAProbes}
\title{Count number of missing values per probe }
\description{Calculates how many missing values there are.}
\usage{ NAProbes(x, latex = FALSE) }
\arguments{
  \item{x}{ a MethIll object}
  \item{latex}{ whether to report the output as latex}
}
\author{ Fabrice Colas}
